+++ assets/javascripts/kitten/components/icons/burger-icon.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-export const BurgerIcon = props => {
-  <svg xmlns="http://www.w3.org/2000/svg"
-       viewBox="0 0 100 125">
-    <path d="M22 29h55.9v9.5H22zM22 61.5h55.9V71H22zM22 45.2h55.9v9.5H22z"/>
-  </svg>
-}